+++ aegis_gym/ros/robot_commander_ros.py
@@ -60,6 +60,9 @@ class RobotCommanderROS(RobotCommanderInterface):
             ],
         )
 
+    def get_base_position(self) -> th.Tensor:
+        return th.zeros(3, dtype=th.float32, device=self.device)
+
     def control_dofs_position(
         self, target_pos: th.Tensor, max_vel: float = 0.3, max_accel: float = 0.3
     ) -> None: